+++ packages/mml-react-types/src/createAttributeGroupsDefinitions.ts
@@ -1,72 +1,96 @@
-import ts from "typescript";
+import { Attribute, AttributeGroup } from "@mml-io/mml-schema";
+import {
+  factory,
+  InterfaceDeclaration,
+  MethodSignature,
+  PropertySignature,
+  SyntaxKind,
+} from "typescript";
 
 import {
-  AttributeGroup,
-  AttributeGroups,
-  createEventHandlersDeclarations,
+  createEventHandlerDeclarations,
   createEventMapInterfaceDeclaration,
-  getAttributeGroupName,
-  getReturnAttributeType,
-} from "./buildDeclarationFile.ts";
+} from "./createEventHandlerDeclarations";
+import { getAttributeGroupName, getReturnAttributeType } from "./util";
 
-const { factory, NodeFlags, NewLineKind, createPrinter, SyntaxKind } = ts;
+export function createAttributeGroupsDefinitions(attributeGroups: {
+  [key: string]: AttributeGroup;
+}): Array<InterfaceDeclaration> {
+  const attributeGroupDeclarations: Array<InterfaceDeclaration> = [];
 
-export function createAttributeGroupsDefinitions(attributes: AttributeGroups) {
-  const attributeGroups = [];
-
-  for (const attributeGroupName in attributes) {
-    const attributeGroup = attributes[attributeGroupName];
+  for (const attributeGroupName in attributeGroups) {
+    const attributeGroup = attributeGroups[attributeGroupName];
     const [instanceDeclaration, attributeDeclaration, eventMapInterfaceDeclaration] =
       createAttributeGroupDeclarations(attributeGroupName, attributeGroup);
-    attributeGroups.push(instanceDeclaration, attributeDeclaration);
+    attributeGroupDeclarations.push(instanceDeclaration, attributeDeclaration);
     if (eventMapInterfaceDeclaration) {
-      attributeGroups.push(eventMapInterfaceDeclaration);
+      attributeGroupDeclarations.push(eventMapInterfaceDeclaration);
     }
   }
 
-  return attributeGroups;
+  return attributeGroupDeclarations;
 }
 
 function createAttributeGroupDeclarations(
   attributeGroupName: string,
   attributeGroup: AttributeGroup,
-) {
+): [InterfaceDeclaration, InterfaceDeclaration, InterfaceDeclaration?] {
   const capitalizedTypeName = getAttributeGroupName(attributeGroupName);
   const typeName = factory.createIdentifier(capitalizedTypeName);
 
-  const attributesWithoutScripts = (attributeGroup.attributes || []).filter(
-    (attribute) => attribute.type !== "Script",
-  );
+  const scriptAttributes: Array<Attribute> = [];
+  const directAttributes: Array<PropertySignature> = [];
+  for (const attribute of attributeGroup.attributes) {
+    if (attribute.type === "Script") {
+      if (attribute.name === "onclick") {
+        directAttributes.push(
+          factory.createPropertySignature(
+            undefined,
+            factory.createIdentifier("onClick"),
+            factory.createToken(SyntaxKind.QuestionToken),
+            factory.createFunctionTypeNode(
+              undefined,
+              [
+                factory.createParameterDeclaration(
+                  undefined,
+                  undefined,
+                  "event",
+                  undefined,
+                  factory.createTypeReferenceNode("MMLClickEvent", undefined),
+                ),
+              ],
+              factory.createKeywordTypeNode(SyntaxKind.VoidKeyword),
+            ),
+          ),
+        );
+      } else {
+        scriptAttributes.push(attribute);
+      }
+    } else {
+      directAttributes.push(
+        factory.createPropertySignature(
+          undefined,
+          factory.createStringLiteral(attribute.name),
+          factory.createToken(SyntaxKind.QuestionToken),
+          factory.createTypeReferenceNode(
+            getReturnAttributeType(attribute),
+            /*typeArguments*/ undefined,
+          ),
+        ),
+      );
+    }
+  }
 
-  const attributeGroupAttributes: any[] = (attributesWithoutScripts || []).map((attribute) => {
-    const type = getReturnAttributeType(attribute);
-
-    const typeNode = factory.createTypeReferenceNode(type, /*typeArguments*/ undefined);
-
-    return factory.createPropertySignature(
-      undefined,
-      // can we create this with the factory methods?
-      factory.createStringLiteral(attribute.name),
-      factory.createToken(SyntaxKind.QuestionToken),
-      typeNode,
-    );
-  });
-
-  let eventHandlersDeclarations: ts.MethodSignature[] = [];
-
-  const eventMapTypeName = factory.createIdentifier(`${capitalizedTypeName}EventMap`);
-
+  let eventHandlersDeclarations: Array<MethodSignature> = [];
   let eventMapInterfaceDeclaration;
-
-  const hasScriptAttributes = attributesWithoutScripts.length !== attributeGroup.attributes?.length;
-
-  if (hasScriptAttributes) {
+  if (scriptAttributes.length > 0) {
+    const eventMapTypeName = factory.createIdentifier(`${capitalizedTypeName}EventMap`);
     eventMapInterfaceDeclaration = createEventMapInterfaceDeclaration(
       eventMapTypeName,
-      attributeGroup.attributes ?? [],
+      scriptAttributes,
     );
 
-    eventHandlersDeclarations = createEventHandlersDeclarations(eventMapTypeName);
+    eventHandlersDeclarations = createEventHandlerDeclarations(eventMapTypeName);
   }
 
   const instanceDeclaration = factory.createInterfaceDeclaration(
@@ -82,7 +106,7 @@ function createAttributeGroupDeclarations(
     factory.createIdentifier(`${capitalizedTypeName}Attributes`),
     undefined,
     undefined,
-    attributeGroupAttributes,
+    directAttributes,
   );
 
   return [instanceDeclaration, attributeDeclaration, eventMapInterfaceDeclaration];
